Check non-null before

+++ drvrgsiftp.c
@@ -217,8 +217,9 @@ int gsiftp_write(int hdl, void *buffer, long nbytes)
 int gsiftp_close(int handle)
 {
     unlink(gsiftp_tmpfile);
-    
-    rmdir(gsiftp_tmpdir);
+
+    if (gsiftp_tmpdir[0] != '\0')
+      rmdir(gsiftp_tmpdir);
 
     return file_close(handle);
 }
